Trigger demo variant modal via button instead of link in info card
