+++ src/components/DemoInfoCard/DemoInfoCard.tsx
@@ -1,34 +1,9 @@
 "use client";
 
 import { InformationSquareIcon } from "@navikt/aksel-icons";
-import { BodyLong, Box, InfoCard, Link } from "@navikt/ds-react";
-import { useState } from "react";
-import DemoVariantPickerModal from "@/components/DemoInfoCard/DemoVariantPickerModal";
-import type { FormVariant } from "@/forms/kartleggingssporsmal/formVariants/types/FormVariant";
-import {
-  DEFAULT_DEMO_FORM_VARIANT,
-  useDemoFormVariantViaParamIfDemo,
-} from "./useDemoFormVariant";
+import { BodyLong, Box, InfoCard } from "@navikt/ds-react";
 
 export default function DemoInfoCard() {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const { activeFormVariant, changeDemoFormVariantViaParam } =
-    useDemoFormVariantViaParamIfDemo(DEFAULT_DEMO_FORM_VARIANT);
-
-  function openModal() {
-    setIsModalOpen(true);
-  }
-
-  function closeModal() {
-    setIsModalOpen(false);
-  }
-
-  function handleChangeDemoFormVariant(variant: FormVariant) {
-    changeDemoFormVariantViaParam(variant);
-    closeModal();
-  }
-
   return (
     <Box paddingBlock="space-24">
       <InfoCard data-color="info" size="small">
@@ -41,27 +16,11 @@ export default function DemoInfoCard() {
           </BodyLong>
 
           <BodyLong size="small">
-            Ulike brukere kan få ulike varianter av skjemaet.{" "}
-            <Link
-              href="#"
-              onClick={(event) => {
-                event.preventDefault();
-                openModal();
-              }}
-            >
-              Velg skjemavariant
-            </Link>{" "}
-            som vises.
+            Skjemavariant kan endres ved å klikke Demo-knappen nederst til
+            høyre.
           </BodyLong>
         </InfoCard.Message>
       </InfoCard>
-
-      <DemoVariantPickerModal
-        open={isModalOpen}
-        activeVariant={activeFormVariant}
-        onClose={closeModal}
-        onSelectVariant={handleChangeDemoFormVariant}
-      />
     </Box>
   );
 }
